lab lifting DONE
